palette-ml/pyproject.toml
[tool.poetry]
name = "palette-ml"
version = "0.1.0"
description = ""
authors = ["Sebastien Rosset"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "<2.0"
pillow = "^11.1.0"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
